Fixed drawing glitch with overwide tiles
The layer rendering code was not prepared to handle tiles that were
wider than the tile width of the map.

This commit also fixes the initialization of the maximum tile height,
which was based on the map height rather than the tile height. This
could slightly reduce overdraw for some maps.

Reviewed-by: Stefan Beller <stefanbeller@googlemail.com>